:recycle: api: move finalized-session-repository to src
